Add quick search in workspace feature
This patch adds a "quick" search in workspace feature.  It works using
the "quick open" widget.  The default keybinding to open it is
ctrl-shift-F.  The user types in the widget and the search results
appear as they type.  Selecting a search result opens the file at the
location of the file.

The backend uses ripgrep to implement the search.  There is no way to
make ripgrep output a truly unambiguous machine parsable output.  The
best way that people have found is to enable the color output, and use
the ANSI color codes as markers for the fields and matches.

The vision in the future would be to have a proper "search in
directory/workspace" view, which would show the results more
permanently.  I think this quick search in workspace feature is still
useful in the mean time, easy to implement on the UI side, and lays out
the foundation in the backend.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>